+++ src/pages/Event/Event.scss
